+++ packages/flutter_cubit/test/cubit_builder_test.dart
@@ -94,7 +94,7 @@ class MyCounterAppState extends State<MyCounterApp> {
           children: <Widget>[
             CubitBuilder<CounterCubit, int>(
               cubit: _cubit,
-              condition: (previousState, state) {
+              buildWhen: (previousState, state) {
                 return (previousState + state) % 3 == 0;
               },
               builder: (context, count) {
@@ -336,7 +336,7 @@ void main() {
       expect(numBuilds, 1);
     });
 
-    testWidgets('with condition only rebuilds when condition evaluates to true',
+    testWidgets('with buildWhen only rebuilds when buildWhen evaluates to true',
         (tester) async {
       await tester.pumpWidget(MyCounterApp());
       await tester.pumpAndSettle();