+++ packages/flutter_cubit/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.1.0
+
+- **BREAKING**: rename `condition` in `CubitBuilder` to `buildWhen`
+- **BREAKING**: rename `condition` in `CubitListener` to `listenWhen`
+- feat: upgrade to `cubit: ^0.1.0`
+- docs: inline documentation improvements
+
 # 0.0.12
 
 - feat: upgrade to `cubit: ^0.0.13`